--- tests/lib/API/HasUserCriterionTest.php
@@ -1,73 +1,0 @@
-<?php
-
-namespace Netgen\EzPlatformSearchExtra\Tests\API;
-
-use eZ\Publish\API\Repository\Values\Content\LocationQuery;
-use eZ\Publish\API\Repository\Values\Content\Query;
-use eZ\Publish\API\Repository\Values\Content\Query\Criterion\ContentId;
-use eZ\Publish\API\Repository\Values\Content\Query\Criterion\LogicalAnd;
-use eZ\Publish\API\Repository\Values\Content\Query\SortClause\ContentId as ContentIdSortClause;
-use Netgen\EzPlatformSearchExtra\API\Values\Content\Query\Criterion\HasUser;
-
-class HasUserCriterionTest extends BaseTest
-{
-    public function providerForTestFind()
-    {
-        return [
-            [
-                new LocationQuery([
-                    'filter' => new LogicalAnd([
-                        new ContentId([10, 14, 41]),
-                        new HasUser(true),
-                    ]),
-                    'sortClauses' => [new ContentIdSortClause()],
-                ]),
-                [10, 14],
-            ],
-            [
-                new LocationQuery([
-                    'filter' => new LogicalAnd([
-                        new ContentId([10, 14, 41]),
-                        new HasUser(false),
-                    ]),
-                    'sortClauses' => [new ContentIdSortClause()],
-                ]),
-                [41],
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider providerForTestFind
-     *
-     * @param \eZ\Publish\API\Repository\Values\Content\Query $query
-     * @param array $expectedIds
-     *
-     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException
-     */
-    public function testFindContent(Query $query, array $expectedIds)
-    {
-        $searchService = $this->getSearchService();
-
-        $searchResult = $searchService->findContentInfo($query);
-
-        $this->assertSearchResultContentIds($searchResult, $expectedIds);
-    }
-
-    /**
-     * @dataProvider providerForTestFind
-     *
-     * @param \eZ\Publish\API\Repository\Values\Content\LocationQuery $query
-     * @param array $expectedIds
-     *
-     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException
-     */
-    public function testFindLocations(LocationQuery $query, $expectedIds)
-    {
-        $searchService = $this->getSearchService();
-
-        $searchResult = $searchService->findLocations($query);
-
-        $this->assertSearchResultContentIds($searchResult, $expectedIds);
-    }
-}
