+++ tests/lib/Integration/API/UserEnabledCriterionTest.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Netgen\EzPlatformSearchExtra\Tests\Integration\API;
+
+use eZ\Publish\API\Repository\Values\Content\LocationQuery;
+use eZ\Publish\API\Repository\Values\Content\Query;
+use eZ\Publish\API\Repository\Values\Content\Query\Criterion\ContentId;
+use eZ\Publish\API\Repository\Values\Content\Query\Criterion\LogicalAnd;
+use eZ\Publish\API\Repository\Values\Content\Query\SortClause\ContentId as ContentIdSortClause;
+use Netgen\EzPlatformSearchExtra\API\Values\Content\Query\Criterion\UserEnabled;
+
+class UserEnabledCriterionTest extends BaseTest
+{
+    public function providerForTestFind()
+    {
+        return [
+            [
+                new LocationQuery([
+                    'filter' => new LogicalAnd([
+                        new ContentId([10, 14, 41]),
+                        new UserEnabled(true),
+                    ]),
+                    'sortClauses' => [new ContentIdSortClause()],
+                ]),
+                [10, 14],
+            ],
+            [
+                new LocationQuery([
+                    'filter' => new LogicalAnd([
+                        new ContentId([10, 14, 41]),
+                        new UserEnabled(false),
+                    ]),
+                    'sortClauses' => [new ContentIdSortClause()],
+                ]),
+                [],
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider providerForTestFind
+     *
+     * @param \eZ\Publish\API\Repository\Values\Content\Query $query
+     * @param array $expectedIds
+     *
+     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException
+     */
+    public function testFindContent(Query $query, array $expectedIds)
+    {
+        $searchService = $this->getSearchService();
+
+        $searchResult = $searchService->findContentInfo($query);
+
+        $this->assertSearchResultContentIds($searchResult, $expectedIds);
+    }
+
+    /**
+     * @dataProvider providerForTestFind
+     *
+     * @param \eZ\Publish\API\Repository\Values\Content\LocationQuery $query
+     * @param array $expectedIds
+     *
+     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException
+     */
+    public function testFindLocations(LocationQuery $query, $expectedIds)
+    {
+        $searchService = $this->getSearchService();
+
+        $searchResult = $searchService->findLocations($query);
+
+        $this->assertSearchResultContentIds($searchResult, $expectedIds);
+    }
+}
