--- tests/lib/Resources/config/services.yml
@@ -1,21 +1,0 @@
-services:
-    netgen_test.search.solr.subdocument_mapper.content:
-        class: Netgen\EzPlatformSearchExtra\Tests\Implementation\Solr\SubdocumentMapper\TestContentSubdocumentMapper
-        tags:
-            - {name: netgen.search.solr.subdocument_mapper.content}
-
-    netgen_test.search.solr.subdocument_mapper.content_translation:
-        class: Netgen\EzPlatformSearchExtra\Tests\Implementation\Solr\SubdocumentMapper\TestContentTranslationSubdocumentMapper
-        tags:
-            - {name: netgen.search.solr.subdocument_mapper.content_translation}
-
-    netgen_test.search.solr.subdocument_mapper.content.sort:
-        class: Netgen\EzPlatformSearchExtra\Tests\Implementation\Solr\SubdocumentMapper\TestSortContentSubdocumentMapper
-        tags:
-            - {name: netgen.search.solr.subdocument_mapper.content}
-
-    netgen_test.search.solr.slot.subdocument_publish_version:
-        parent: ezpublish.search.solr.slot
-        class: '%ezpublish.search.solr.slot.publish_version.class%'
-        tags:
-            - {name: ezpublish.search.solr.slot, signal: ContentService\PublishVersionSignal}
